requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
discord.py==1.5.1
